backend/requirement.txt
fastapi
uvicorn
pydantic
python-multipart
langchain
langchain-community
faiss-cpu
openai
wikipedia
python-docx
docx2pdf
httpx
langgraph
python-dotenv
google-generativeai
pypdf
unstructured[excel]
docx2txt
pandas
docx
langchain_google_genai